+++ src/main/java/com/xiaoleilu/hutool/convert/impl/DoubleArrayConverter.java
@@ -1,0 +1,20 @@
+package com.xiaoleilu.hutool.convert.impl;
+
+import com.xiaoleilu.hutool.convert.AbstractConverter;
+import com.xiaoleilu.hutool.convert.Convert;
+import com.xiaoleilu.hutool.util.CollectionUtil;
+
+/**
+ * double 类型数组转换器
+ * @author Looly
+ *
+ */
+public class DoubleArrayConverter extends AbstractConverter<double[]>{
+	
+	@Override
+	protected double[] convertInternal(Object value) {
+		final Double[] result = Convert.convert(Double[].class, value);
+		return CollectionUtil.unWrap(result);
+	}
+
+}