+++ src/main/java/com/xiaoleilu/hutool/convert/impl/TimeZoneConverter.java
@@ -1,0 +1,19 @@
+package com.xiaoleilu.hutool.convert.impl;
+
+import java.util.TimeZone;
+
+import com.xiaoleilu.hutool.convert.AbstractConverter;
+
+/**
+ * TimeZone转换器
+ * @author Looly
+ *
+ */
+public class TimeZoneConverter extends AbstractConverter<TimeZone>{
+
+	@Override
+	protected TimeZone convertInternal(Object value) {
+		return TimeZone.getTimeZone(convertToStr(value));
+	}
+
+}