+++ src/main/java/com/xiaoleilu/hutool/util/BeetlUtil.java
@@ -1,0 +1,13 @@
+package com.xiaoleilu.hutool.util;
+
+/**
+ * Beetl模板引擎工具类<br>
+ * http://git.oschina.net/xiandafu/beetl2.0
+ * 文档：http://ibeetl.com/guide/beetl.html
+ * 
+ * @author Looly
+ */
+public class BeetlUtil {
+	public static void render(){
+	}
+}